refactor: Create debug package in o.o.ext.interactivelauncher
