+++ src/ext/java/org/opentripplanner/ext/interactivelauncher/debug/logging/OTPDebugLoggers.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.ext.interactivelauncher.logging;
+package org.opentripplanner.ext.interactivelauncher.debug.logging;
 
 import java.util.List;
